How to become a Hacker
Hacking is an engaging field but it is surely not easy. To become a hacker one has to have an attitude and curiosity of learning and adapting new skills. 
You must have a deep knowledge of computer systems, programming languages, operating systems and the journey of learning goes on and on.


Some people think that a hacker is always a criminal and do illegal activities but they are wrong. Actually many big companies hire hackers to protect their 
systems and information and are highly paid. here is the list of most important steps necessary to become a hacker, have a deeper look